fix: harden usage meters and iOS composer
